Fix CRS download, modularize
