+++ code/3.0_virtual_policy_marker_analysis.R
@@ -64,7 +64,7 @@ rotate_x_text_90 = theme(
 
 crs = fread("large_output/crs_2014_2023_phi4_labeled.csv")
 
-crs$`Sector code` = (crs$sector_code %in% c(16030, 16040))
+crs$`Sector code` = (crs$PurposeCode %in% c(16030, 16040))
 crs$any = crs$`Housing general` | crs$Homelessness |
   crs$`Transitional housing` | crs$`Incremental housing` | crs$`Social housing` |
   crs$`Market housing` | crs$`Sector code`
