+++ output/sector_Year.csv
@@ -1,0 +1,11 @@
+Year,Housing policy and administrative management,Low-cost housing
+2014,230.576004,703.38143
+2015,598.0885,518.323393
+2016,354.863361,522.859276
+2017,488.668093,802.167095
+2018,591.872359,355.29013
+2019,1537.080369,578.738318
+2020,1766.954544,502.951339
+2021,581.574094,377.633947
+2022,730.315401000001,531.20167
+2023,481.469684,333.833324
