+++ output/sector_by_donor_type.csv
@@ -1,5 +1,5 @@
 donor_type,value
-Multilateral,9714.62115600002
-DAC,1708.535457
-Non-DAC,452.73471
-Private,29.824956
+Multilateral,10239.21746
+DAC,1830.96223500001
+Non-DAC,476.100212
+Private,41.562424
