+++ output/sector_by_income.csv
@@ -1,7 +1,7 @@
-income_group_code,value
-LMICs,5614.461843
-UMICs,5294.95456
-LDCs,590.959026
-Part I unallocated by income,341.523139
-MADCTs,55.928279
-Other LICs,7.889432
+IncomegroupName,value
+UMICs,6289.843983
+LMICs,5273.008783
+LDCs,593.644478
+Part I unallocated by income,364.136825
+MADCTs,58.63966
+Other LICs,8.568602
