+++ output/sector_category_Year.csv
@@ -1,0 +1,11 @@
+Year,ODA,OOF,Private
+2014,283.236531,649.977044,0.743859
+2015,331.182781,784.389708,0.839404
+2016,370.930406,504.645266,2.146965
+2017,626.300829999999,656.3597,8.174658
+2018,396.210794,546.319853,4.631842
+2019,590.063265,1518.178,7.577422
+2020,473.850568,1792.50534,3.549975
+2021,193.373539,763.144768,2.689734
+2022,244.504761,1015.903745,1.108565
+2023,194.389197,532.186989,88.726822
