--- output/sector_category_year.csv
@@ -1,11 +1,0 @@
-year,ODA,OOF,Private
-2014,265.819998,621.503808,0.712874
-2015,310.073139,749.409626,0.772996
-2016,350.07006,480.405993,2.041958
-2017,581.985478,624.933067,7.749542
-2018,372.008704,519.807648,4.441862
-2019,557.043863,1444.084605,7.268529
-2020,451.49774,1708.340234,3.412551
-2021,178.70199,698.930706,2.557163
-2022,229.535336,962.571676,0.867481
-2023,184.803379,509.648243,74.71603
