--- output/sector_year.csv
@@ -1,11 +1,0 @@
-year,Housing policy and administrative management,Low-cost housing
-2014,220.590817,667.445863
-2015,569.004284999999,491.251476
-2016,337.430898,495.087113
-2017,464.40998,750.258107
-2018,560.372002,335.886212
-2019,1460.276904,548.120093
-2020,1681.50942,481.741105
-2021,523.584167999999,356.605691
-2022,690.545155,502.429338000001
-2023,449.375777,319.791874999999
