+++ requirements.txt
@@ -12,6 +12,7 @@ attrs==24.3.0
 babel==2.16.0
 beautifulsoup4==4.12.3
 bleach==6.2.0
+cachetools==5.5.1
 certifi==2024.12.14
 cffi==1.17.1
 charset-normalizer==3.4.1
@@ -36,9 +37,21 @@ fonttools==4.55.7
 fqdn==1.5.1
 frozenlist==1.5.0
 fsspec==2024.9.0
+google==3.0.0
+google-ai-generativelanguage==0.6.15
+google-api-core==2.24.1
+google-api-python-client==2.160.0
+google-auth==2.38.0
+google-auth-httplib2==0.2.0
+google-genai==1.0.0
+google-generativeai==0.8.4
+googleapis-common-protos==1.66.0
 greenlet==3.1.1
+grpcio==1.70.0
+grpcio-status==1.70.0
 h11==0.14.0
 httpcore==1.0.7
+httplib2==0.22.0
 httpx==0.28.1
 huggingface-hub==0.27.0
 idna==3.10
@@ -98,7 +111,7 @@ nvidia-cusparse-cu12==12.3.1.170
 nvidia-nccl-cu12==2.21.5
 nvidia-nvjitlink-cu12==12.4.127
 nvidia-nvtx-cu12==12.4.127
-oda_reader==1.0.0
+oda_reader==1.1.5
 ollama==0.4.7
 openai==1.58.1
 openai-function-tokens==0.1.2
@@ -114,11 +127,15 @@ platformdirs==4.3.6
 prometheus_client==0.21.1
 prompt_toolkit==3.0.48
 propcache==0.2.1
+proto-plus==1.26.0
+protobuf==5.29.3
 psutil==6.1.1
 psycopg2-binary==2.9.10
 ptyprocess==0.7.0
 pure_eval==0.2.3
 pyarrow==18.1.0
+pyasn1==0.6.1
+pyasn1_modules==0.4.1
 pycparser==2.22
 pydantic==2.10.4
 pydantic_core==2.27.2
@@ -136,6 +153,7 @@ requests==2.32.3
 rfc3339-validator==0.1.4
 rfc3986-validator==0.1.1
 rpds-py==0.22.3
+rsa==4.9
 safetensors==0.5.0
 scikit-learn==1.6.1
 scipy==1.15.1
@@ -164,13 +182,16 @@ types-python-dateutil==2.9.0.20241206
 typing_extensions==4.12.2
 tzdata==2024.2
 uri-template==1.3.0
+uritemplate==4.1.1
 urllib3==2.3.0
 wcwidth==0.2.13
 webcolors==24.11.1
 webencodings==0.5.1
 websocket-client==1.8.0
+websockets==14.2
 widgetsnbextension==4.0.13
 xformers==0.0.29.post1
+xgboost==2.1.3
 xmltodict==0.14.2
 xxhash==3.5.0
 yarl==1.18.3
